requirements.txt
Flask==2.2.5
google-cloud-speech==2.16.2
google-api-python-client==2.70.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
